+++ src/main/java/org/deepsymmetry/beatlink/data/WaveformFinder.java
@@ -467,7 +467,7 @@ public class WaveformFinder extends LifecycleParticipant {
     WaveformPreview getWaveformPreview(int rekordboxId, SlotReference slot, Client client)
             throws IOException {
         Message response = client.simpleRequest(Message.KnownType.WAVE_PREVIEW_REQ, null,
-                client.buildRMS1(Message.MenuIdentifier.DATA, slot.slot), NumberField.WORD_1,
+                client.buildRMST(Message.MenuIdentifier.DATA, slot.slot), NumberField.WORD_1,
                 new NumberField(rekordboxId), NumberField.WORD_0);
         if (response.knownType == Message.KnownType.WAVE_PREVIEW) {
             return new WaveformPreview(new DataReference(slot, rekordboxId), response);
@@ -581,7 +581,7 @@ public class WaveformFinder extends LifecycleParticipant {
     WaveformDetail getWaveformDetail(int rekordboxId, SlotReference slot, Client client)
             throws IOException {
         Message response = client.simpleRequest(Message.KnownType.WAVE_DETAIL_REQ, null,
-                client.buildRMS1(Message.MenuIdentifier.MAIN_MENU, slot.slot),
+                client.buildRMST(Message.MenuIdentifier.MAIN_MENU, slot.slot),
                 new NumberField(rekordboxId), NumberField.WORD_0);
         if (response.knownType == Message.KnownType.WAVE_DETAIL) {
             return new WaveformDetail(new DataReference(slot, rekordboxId), response);